Updated todolist, separated app into components
